Bug 1210912 - fix widget leak in LookupRegisteredPluginWindow; r=roc
nsRefPtrHashtable::Get follows XPCOM outparam rules, so it addrefs its
argument prior to returning it.  LookupRegisteredPluginWindow, however,
doesn't know anything about the additional reference, and neither does
anything else that calls it (which is, fortunately, not that much code).
So every widget returned by this function will leak.

To fix this, we use nsRefPtrHashtable::GetWeak, which doesn't addref its
argument, but merely returns the raw pointer, leaving the hashtable
responsible for the owning reference.