Bug 1225923 - part 2 - fix AppendElement(nsDependentString(...)) call in HTMLInputElement.cpp; r=baku
I think the intent of this call is to not copy the filename data passed
in, but to simply convert it to a friendlier nsString container for
processing.  We don't have to worry about the nsTArray<nsString>
overload of MozSetFileNameArray holding references to the strings
outside the call, as the elements of the array are copied into new
strings as appropriate.