Bug 1233863 - ARM64: Record the correct branch offset. r=sstangl
Don't depend on nextOffset() to get the address of the next inserted
instruction. Adding a single instruction could cause a constant pool to be
emitted first.

The b() method assembles a branch and returns its offset. Use this return value
when recording the location of a pending jump.

This fixes the MOZ_CRASH(Unrecognized jump instruction.) crashes.