Bug 1233111 - Implement SIMD shiftRightByScalar(). r=bbouvie
This is the right shift function that the SIMD.js spec requires. The old
shiftRightArithmeticByScalar() and shiftRightLogicalByScalar() functions will go
away.

These functions perform an arithmetic shift for signed types and a logical
shift for unsigned types.

Add support to Odin and Ion too, at least for the Int32x4 variant.