Bug 1244889 - Define a canonical subset of Type::Which. r=luke
This subset of the asm.js types corresponds to the Void type and all the
coercion targets. It corresponds to wasm::ExprType, except:

- asm.js does not support I64.
- asm.js distinguishes signed/unsigned SIMD types, wasm does not.

The function Type::canonicalize() maps a type into the canonical set.
Type::isCanonical() checks for membership.

This subset of Type::Which will be used to replace wasm::ValType and
wasm::ExprType in the asm.js type checking code. This makes it possible to
type-check unsigned SIMD types correctly.

Eliminate Type:checkedValueType(). Use Type::canonicalToValType() instead.