Bug 1236053 - Remove the filtering functionality from the Tree component; r=jsantell
In practice, we filter trees off-main-thread in a worker and the tree that we
pass to the Tree component is already filtered, which is a better architecture
anyways. Additionally, we lack any kind of tests for the Tree component's
built-in filtering. And finally, there isn't even a single user of the built-in
filtering functionality.