Bug 1198448 - The Firefox UI update tests should call a different entry point. NPOTB. DONTBUILD. r=chmanchester
When setting up the Firefox UI tests repository, it generates two binaries, one of them
called firefox-ui-update. Calling this binary on Windows triggers the UAC prompt for the
Release Engineering Windows test machines.

If instead we use Python to call a script that calls the same functionality we don't
get this UAC prompt.

We also move around some pre action checks to the right place. This allows us to not need to
specify --firefox-ui-branch when we're not running the checkout action.