Bug 1141041 - Stop using CPOWs for the Page Style menu. r=Gijs
Instead of using a CPOW to synchronously grab the stylesheet information
from the currently loaded tab, each tab now sends up the stylesheet
information once they've finished loading for the parent to cache.

Unfortunately, the cache will not be invalidated if the stylesheets on
a page are dynamically altered with script.